+++ history.md
@@ -7,6 +7,7 @@
 - cookies set in a response should be kept in a redirect
 - HEAD requests should process parameters just like GET (patch provided by Rob Eanes)
 - Exception message should never be nil (patch provided by Michael Klett)
+- Limit number of redirections (patch provided by Chris Dinn)
 
 # 1.6.0